Move getImportSource to import-sources
